added changes to backend model fo better adaption
