+++ backend/apiauth/admin.py
@@ -1,4 +1,5 @@
 from django.contrib import admin
-from .models import UserProfile
+from .models import UserProfile, TechStack
 
 admin.site.register(UserProfile)
+admin.site.register(TechStack)
