+++ backend/apiauth/models.py
@@ -1,6 +1,19 @@
 from django.db import models
 from django.contrib.auth.models import User
 
+
+class TechStack(models.Model):
+    #Model
+    tech_name = models.CharField(max_length=60)
+
+    class Meta:
+        verbose_name = 'Tech Stack'
+        verbose_name_plural = 'Tech Stack'
+
+    def __str__(self):
+        return self.tech_name
+
+
 class UserProfile(models.Model):
     PROG_CHOICES = (
         ('BT', 'B.Tech'),
@@ -39,8 +52,12 @@ class UserProfile(models.Model):
     phone = models.CharField(max_length=10, null=True)
     prog = models.CharField(max_length=5, choices=PROG_CHOICES, verbose_name='Programme', default='BT')
     year = models.CharField(max_length=1, choices=YEAR_CHOICES, default='1')
-    avatar = models.ImageField(upload_to='avatar', default='default_member.png', height_field=None, width_field=None)
+    avatar = models.ImageField(upload_to='avatar', default='default_member.png', height_field=None, width_field=None, blank=True)
     branch = models.CharField(max_length=5, choices=BRANCH_CHOICES)
+    techstack = models.ManyToManyField(TechStack, blank=True)
+    github = models.URLField(max_length=200, null=True, blank=True)
+    linkedin = models.URLField(max_length=200, null=True, blank=True)
+
     
     def __str__(self):
         return self.user.get_full_name()
